[update]CustomAdapterを独立させる
